refactor(checker): split TypeFormatter factories out of def_mapping.rs (2000-line cap)
The lazy-DefId-cache change pushed def_mapping.rs to 2019 lines, over the 2000
physical-line cap (checker-boundary arch guard). Extract the three formatter
construction factories (create_type_formatter / create_diagnostic_type_formatter
/ create_assignability_type_formatter) into a new def_mapping_formatters.rs
sibling module on the same `impl<'a> CheckerContext<'a>`. def_mapping.rs → 1942
lines; no behavior change (functions moved verbatim).
