+++ crates/tsz-checker/src/context/mod.rs
@@ -34,6 +34,7 @@ mod file_session_reset;
 pub mod lifetime_shells;
 pub use lifetime_shells::{FileSession, LspPersistentCache, SpeculationScope, WorkerContext};
 mod def_mapping;
+mod def_mapping_formatters;
 mod deferred_flow_env_write;
 pub use deferred_flow_env_write::DeferredFlowEnvWrite;
 mod file_format_lookup;
